+++ src/components/Explorer/ExplorerFormField.jsx
@@ -1,11 +1,14 @@
 import React from 'react';
 import AutoComplete from 'material-ui/AutoComplete';
 import DatePicker from 'material-ui/DatePicker';
+// import FormField from 'components/Form/FormField';
 
 class ExplorerFormField extends React.Component {
   constructor() {
     super();
     this.resetField = this.resetField.bind(this);
+    this.addChip = this.addChip.bind(this);
+    this.deleteChip = this.deleteChip.bind(this);
   }
   resetField() {
     const { builderField, handleFieldUpdate } = this.props;
@@ -19,6 +22,19 @@ class ExplorerFormField extends React.Component {
       this.datepicker.setState({ date: undefined });
     }
     handleFieldUpdate(builderField, undefined);
+  }
+  addChip(name, input, limit) {
+    const currentChips = [].concat(this.props.builder[name] || []);
+    const newChips = [input.key].concat(currentChips).slice(0, limit);
+    this.props.handleFieldUpdate(name, newChips);
+  }
+  deleteChip(name, index) {
+    const currentChips = [].concat(this.props.builder[name] || []);
+    const newChips = [
+      ...currentChips.slice(0, index),
+      ...currentChips.slice(index + 1),
+    ];
+    this.props.handleFieldUpdate(name, newChips);
   }
   render() {
     const { fields, label, builderField, handleFieldUpdate, isDateField, builder } = this.props;
@@ -39,6 +55,19 @@ class ExplorerFormField extends React.Component {
         /></span>);
     }
     return (<span style={{ width: fieldWidth }}>
+      {/*
+      <FormField
+        name={builderField}
+        label={label}
+        dataSource={dataSource}
+        formSelectionState={builder}
+        filter={AutoComplete.caseInsensitiveFilter}
+        strict
+        limit={1}
+        addChip={this.addChip}
+        deleteChip={this.deleteChip}
+      />
+      */}
       <AutoComplete
         ref={ref => (this.autocomplete = ref)}
         searchText={builder[builderField]
