+++ src/components/Player/TableFilterForm/index.jsx
@@ -6,8 +6,7 @@ import { withRouter } from 'react-router-dom';
 import querystring from 'querystring';
 import strings from 'lang';
 import { toggleShowForm } from 'actions/formActions';
-import Form from 'components/Form/Form';
-import FormGroup from 'components/Form/FormGroup';
+import FormField from 'components/Form/FormField';
 import styles from './TableFilterForm.css';
 import * as data from './TableFilter.config';
 
@@ -52,109 +51,131 @@ class TableFilterForm extends React.Component {
     return (
       <div>
         <div className={showForm ? styles.showForm : styles.hideForm}>
-          <Form name={FORM_NAME} className={styles.form}>
-            <FormGroup formSelectionState={formSelectionState} history={history}>
-              {Field => (
-                <div className={styles.formGroup}>
-                  <Field
-                    name="hero_id"
-                    label={strings.filter_hero_id}
-                    dataSource={data.heroList}
-                    strict
-                    limit={1}
-                  />
-                  <Field
-                    name="is_radiant"
-                    label={strings.filter_is_radiant}
-                    dataSource={data.factionList}
-                    strict
-                    limit={1}
-                  />
-                  <Field
-                    name="win"
-                    label={strings.filter_win}
-                    dataSource={data.resultList}
-                    strict
-                    limit={1}
-                  />
-                  <Field
-                    name="lane_role"
-                    label={strings.filter_lane_role}
-                    dataSource={data.laneList}
-                    strict
-                    limit={1}
-                  />
-                  <Field
-                    name="patch"
-                    label={strings.filter_patch}
-                    dataSource={data.patchList}
-                    strict
-                    limit={1}
-                  />
-                  <Field
-                    name="game_mode"
-                    label={strings.filter_game_mode}
-                    dataSource={data.modeList}
-                    strict
-                    limit={1}
-                  />
-                  <Field
-                    name="lobby_type"
-                    label={strings.filter_lobby_type}
-                    dataSource={data.lobbyTypeList}
-                    strict
-                    limit={1}
-                  />
-                  <Field
-                    name="date"
-                    label={strings.filter_date}
-                    dataSource={data.dateList}
-                    strict
-                    limit={1}
-                  />
-                  <Field
-                    name="region"
-                    label={strings.filter_region}
-                    dataSource={data.regionList}
-                    strict
-                    limit={1}
-                  />
-                  <Field
-                    name="with_hero_id"
-                    label={strings.filter_with_hero_id}
-                    dataSource={data.heroList}
-                    strict
-                    limit={5}
-                  />
-                  <Field
-                    name="against_hero_id"
-                    label={strings.filter_against_hero_id}
-                    dataSource={data.heroList}
-                    strict
-                    limit={5}
-                  />
-                  <Field
-                    name="included_account_id"
-                    label={strings.filter_included_account_id}
-                    dataSource={this.state.peers.map(peer => ({ text: `${peer.personaname}`, value: peer.account_id }))}
-                    limit={10}
-                  />
-                  <Field
-                    name="excluded_account_id"
-                    label={strings.filter_excluded_account_id}
-                    dataSource={this.state.peers.map(peer => ({ text: `${peer.personaname}`, value: peer.account_id }))}
-                  />
-                  <Field
-                    name="significant"
-                    label={strings.filter_significant}
-                    dataSource={data.significantList}
-                    strict
-                    limit={1}
-                  />
-                </div>
-              )}
-            </FormGroup>
-          </Form>
+          <div className={styles.formGroup}>
+            <FormField
+              name="hero_id"
+              label={strings.filter_hero_id}
+              dataSource={data.heroList}
+              formSelectionState={formSelectionState}
+              history={history}
+              strict
+              limit={1}
+            />
+            <FormField
+              name="is_radiant"
+              label={strings.filter_is_radiant}
+              dataSource={data.factionList}
+              formSelectionState={formSelectionState}
+              history={history}
+              strict
+              limit={1}
+            />
+            <FormField
+              name="win"
+              label={strings.filter_win}
+              dataSource={data.resultList}
+              formSelectionState={formSelectionState}
+              history={history}
+              strict
+              limit={1}
+            />
+            <FormField
+              name="lane_role"
+              label={strings.filter_lane_role}
+              dataSource={data.laneList}
+              formSelectionState={formSelectionState}
+              history={history}
+              strict
+              limit={1}
+            />
+            <FormField
+              name="patch"
+              label={strings.filter_patch}
+              dataSource={data.patchList}
+              formSelectionState={formSelectionState}
+              history={history}
+              strict
+              limit={1}
+            />
+            <FormField
+              name="game_mode"
+              label={strings.filter_game_mode}
+              dataSource={data.modeList}
+              formSelectionState={formSelectionState}
+              history={history}
+              strict
+              limit={1}
+            />
+            <FormField
+              name="lobby_type"
+              label={strings.filter_lobby_type}
+              dataSource={data.lobbyTypeList}
+              formSelectionState={formSelectionState}
+              history={history}
+              strict
+              limit={1}
+            />
+            <FormField
+              name="date"
+              label={strings.filter_date}
+              dataSource={data.dateList}
+              formSelectionState={formSelectionState}
+              history={history}
+              strict
+              limit={1}
+            />
+            <FormField
+              name="region"
+              label={strings.filter_region}
+              dataSource={data.regionList}
+              formSelectionState={formSelectionState}
+              history={history}
+              strict
+              limit={1}
+            />
+            <FormField
+              name="with_hero_id"
+              label={strings.filter_with_hero_id}
+              dataSource={data.heroList}
+              formSelectionState={formSelectionState}
+              history={history}
+              strict
+              limit={5}
+            />
+            <FormField
+              name="against_hero_id"
+              label={strings.filter_against_hero_id}
+              dataSource={data.heroList}
+              formSelectionState={formSelectionState}
+              history={history}
+              strict
+              limit={5}
+            />
+            <FormField
+              name="included_account_id"
+              label={strings.filter_included_account_id}
+              dataSource={this.state.peers.map(peer => ({ text: `${peer.personaname}`, value: peer.account_id }))}
+              formSelectionState={formSelectionState}
+              history={history}
+              limit={10}
+            />
+            <FormField
+              name="excluded_account_id"
+              label={strings.filter_excluded_account_id}
+              dataSource={this.state.peers.map(peer => ({ text: `${peer.personaname}`, value: peer.account_id }))}
+              formSelectionState={formSelectionState}
+              history={history}
+            />
+            <FormField
+              name="significant"
+              label={strings.filter_significant}
+              dataSource={data.significantList}
+              formSelectionState={formSelectionState}
+              history={history}
+              strict
+              limit={1}
+            />
+          </div>
         </div>
       </div>
     );
